In-progress: encoding using a streaming dataset.
We can use `map` to load and resize the images, as it processes items on
the fly. This makes the code very simple, but we have a bottleneck on
disk I/O and CPU as images are processed sequentially.

I added a simple Pool to parallelize image preparation, and performance
improved a lot, but I'm not convinced it's the right way to do it. More
work needed.