Add "streaming" version of encoding notebook.
The main difference (aside from loading the dataset in streaming mode)
is that images are prepared (resized and cropped) using a `map`
function. This is very convenient, because samples are processed as
needed. In the non-streaming mode of operation, we avoid a similar
`map` operation for these reasons:
- We'd need a lot of space to cache the mapped dataset.
- I found image transformation (CPU-bound), interleaved with
encoding (TPU) to be efficient. Encoding took ~4h for a local SSD copy
of the dataset.

It would be interesting to check how fast it is to pre-process all image
transformations and then run the encoding pass on its own.