Save wandb run and artifact version in text files.
- The run id is reused in subsequent calls so predictions are logged to
the same run.
- The script does nothing if the downloaded artifact contains the same
model version as the previous run.

These changes are intended to run the script periodically (from a cron
job, maybe) while a model is being trained, so results from all model
versions are logged to the same run and not overwritten.

To make it easier to use in other scenarios, we could use command-line
arguments to ignore this behaviour.