[WIP] Attempt to use masked optimizer for adafactor
I tried to adapt the code from
https://colab.research.google.com/drive/1g_pt2Rc3bv6H6qchvGHD-BpgF-Pt4vrC#scrollTo=WWHlukuvIpXb,
but it's hard due to our split and partitioning mechanism.

It currently works (I think) for adafactor, but not for adam or
distributed_shampoo.

The idea we use here is to replace the optimizer with a
`multi_transform` that applies different optimizations to frozen vs
trainable parameters. I believe that optax.masked() does something similar.
I also explored the approach to create a new "frozen" split to zero those gradients, but I had trouble with axis when pjitting the state.